Organise tsumemi module with subfolders (submodules) for settings and board gui
