+++ tsumemi/src/tsumemi/board_gui/komadai_artist.py
