+++ tsumemi/src/tsumemi/kif_browser_gui.py
@@ -14,7 +14,7 @@ import tsumemi.src.tsumemi.event as evt
 import tsumemi.src.tsumemi.game_controller as gamecon
 import tsumemi.src.tsumemi.problem_list as plist
 import tsumemi.src.tsumemi.problem_list_controller as plistcon
-import tsumemi.src.tsumemi.settings_controller as setcon
+import tsumemi.src.tsumemi.settings.settings_controller as setcon
 import tsumemi.src.tsumemi.speedrun_controller as speedcon
 import tsumemi.src.tsumemi.timer as timer
 import tsumemi.src.tsumemi.timer_controller as timecon
