+++ tsumemi/src/tsumemi/settings/piece_setting_choices.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import os
 import tkinter as tk
 
 from tkinter import ttk
@@ -7,31 +8,31 @@ from typing import TYPE_CHECKING
 from PIL import Image, ImageTk
 
 import tsumemi.src.tsumemi.img_handlers as imghand
-import tsumemi.src.tsumemi.setting_choices as setc
+import tsumemi.src.tsumemi.settings.setting_choices as setc
 
 if TYPE_CHECKING:
     from typing import List, Optional
 
 
-class BoardSkinChoice(setc.Choice[imghand.BoardSkin]):
+class PieceSkinChoice(setc.Choice[imghand.PieceSkin]):
     pass
 
 
-BOARD_SKIN_CHOICES: List[BoardSkinChoice] = [
-    BoardSkinChoice(skin, skin.desc, skin.name) for skin in imghand.BoardSkin
+PIECE_SKIN_CHOICES: List[PieceSkinChoice] = [
+    PieceSkinChoice(skin, skin.desc, skin.name) for skin in imghand.PieceSkin
 ]
 
 
-class BoardSkinSelection(setc.Selection[imghand.BoardSkin]):
+class PieceSkinSelection(setc.Selection[imghand.PieceSkin]):
     pass
 
 
-class BoardSkinSelectionController:
+class PieceSkinSelectionController:
     def __init__(self) -> None:
-        self.model = BoardSkinSelection(BOARD_SKIN_CHOICES)
+        self.model = PieceSkinSelection(PIECE_SKIN_CHOICES)
         return
     
-    def get_board_skin(self) -> imghand.BoardSkin:
+    def get_piece_skin(self) -> imghand.PieceSkin:
         return self.model.get_item()
     
     def get_config_string(self) -> str:
@@ -42,20 +43,20 @@ class BoardSkinSelectionController:
         return
 
 
-class BoardSkinDropdown(setc.Dropdown[imghand.BoardSkin]):
+class PieceSkinDropdown(setc.Dropdown[imghand.PieceSkin]):
     pass
 
 
-class BoardSkinSelectionFrame(ttk.Frame):
+class PieceSkinSelectionFrame(ttk.Frame):
     PREVIEW_WIDTH_HEIGHT = (33, 36)
     def __init__(self,
             parent: tk.Widget,
-            controller: BoardSkinSelectionController,
+            controller: PieceSkinSelectionController,
         ) -> None:
         self.controller = controller
         super().__init__(parent)
-        self.lbl_name = ttk.Label(self, text="Board set")
-        self.cmb_dropdown = BoardSkinDropdown(parent=self, controller=controller.model)
+        self.lbl_name = ttk.Label(self, text="Piece set")
+        self.cmb_dropdown = PieceSkinDropdown(parent=self, controller=controller.model)
         # mypy doesn't recognise the "add" parameter overload to bind
         self.cmb_dropdown.bind( # type: ignore
             "<<ComboboxSelected>>", self.set_preview, add="+"
@@ -73,30 +74,15 @@ class BoardSkinSelectionFrame(ttk.Frame):
         return
     
     def set_preview(self, event: Optional[tk.Event]) -> None:
-        skin = self.controller.get_board_skin()
+        skin = self.controller.get_piece_skin()
         filepath = skin.path
         if filepath:
-            img = Image.open(filepath).resize(self.PREVIEW_WIDTH_HEIGHT)
+            filename = os.path.join(filepath, "0GI.png")
+            img = Image.open(filename).resize(self.PREVIEW_WIDTH_HEIGHT)
+            self.lbl_preview["text"] = ""
         else:
-            img = Image.new("RGB", self.PREVIEW_WIDTH_HEIGHT, skin.colour)
-        self.preview_photoimage = ImageTk.PhotoImage(img)
-        self.lbl_preview["image"] = self.preview_photoimage
-        return
-
-
-class KomadaiSkinSelectionFrame(BoardSkinSelectionFrame):
-    def __init__(self,
-            parent: tk.Widget,
-            controller: BoardSkinSelectionController
-        ) -> None:
-        super().__init__(parent=parent, controller=controller)
-        self.lbl_name["text"] = "Komadai"
-        return
-    
-    def set_preview(self, event: Optional[tk.Event]) -> None:
-        skin = self.controller.get_board_skin()
-        filepath = skin.path
-        img = Image.new("RGB", self.PREVIEW_WIDTH_HEIGHT, skin.colour)
+            img = Image.new("RGB", self.PREVIEW_WIDTH_HEIGHT, "#FFFFFF")
+            self.lbl_preview["text"] = "銀"
         self.preview_photoimage = ImageTk.PhotoImage(img)
         self.lbl_preview["image"] = self.preview_photoimage
         return
