+++ tsumemi/src/tsumemi/settings/setting_choices.py
